requirements.txt
Django
django-simple-history

git+https://github.com/botswana-harvard/edc-base@develop#egg=edc-base
git+https://github.com/botswana-harvard/edc-constants@develop#egg=edc-constants
git+https://github.com/botswana-harvard/edc-identifier@develop#egg=edc-identifier
git+https://github.com/botswana-harvard/edc-device@develop#egg=edc-device
git+https://github.com/erikvw/django-revision@develop#egg=django-revision
git+https://github.com/erikvw/django-crypto-fields@develop#egg=django-crypto-fields
